organize components into their own files
